# sqli-server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: SQL Injection Attack
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /v1/login:
    x-swagger-router-controller: login.controller
    post:
      operationId: v1_login
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/login_model"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/login_model"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /v1/isLoggedIn: 
    x-swagger-router-controller: login.controller
    get:
      operationId: v1_isLoggedIn
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DefaultResponse"
        "400":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DefaultResponse"
        "403":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DefaultResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  DefaultResponse:
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  login_model:
      required:
        - email
        - password
      properties:
          email:
            type: string
          password:
              type: string
          query:
            type: string
...